Role now passes molecule test
To pass molecule test the variable 'domains' needs to be set.
While testing noticed, that we need make sure that some packages are
installed.

We also need to prevent Let's Encrypt / dehydrated to sign the keys when
we run this role in molecule test. We currently do this by testing for a
docker environment. This is not the best way to that, but we currently
do not know how to do it better.